refactor: TIO-97: move rhythm widgets in sub folder
