+++ lib/pages/metronome/rhythm/rhythm_generator_setting_list_item.dart
